+++ src/main/java/org/opentripplanner/updater/alert/GtfsRealtimeAlertsUpdater.java
@@ -53,7 +53,7 @@ public class GtfsRealtimeAlertsUpdater extends PollingGraphUpdater implements Tr
 
     LOG.info(
       "Creating real-time alert updater running every {} seconds : {}",
-      pollingPeriodSeconds(),
+      pollingPeriod(),
       url
     );
   }